Create biathlon.cpp
Biathlon competitions consist of jogging and skiing with rifle shooting. Participants must not only run fast, but to shoot accurately. For each missed target is added 30 / thirty / seconds to the time of running. Each participant has to make 5 / five / shots. The winner is the one who takes the least time. Write a program biathlon, which finds and prints the time of the winner if the competition involves three biathlon - A, B and C. There are no competitors with equal achievements.