Create maxdigit.cpp
You are given three positive three-digit integers B, C, D and an positive single-digit integer K. The value of K can only be 1, 2 or 3. Write a program maxdigit, which defines the largest K-digit of B, C, D, which in place of A makes correct inequality 9+ (A-5)> 8 Digits of the number are counted from left to right. If the inequality is not satisfied for any one of those digits, print No.