Create aquarium.cpp
For the normal growth of the goldfish there should be vakvarium at least three liters of water for each fish. Write a program , which determines whether an aquarium with volume V liters is "crowded", if there are N goldfish and, in the case of "Overpopulation" finds the minimum amount of fish that should be moved to another place in order to have the necessary amount of water for the remaining in the aquarium fish.