Create traffic.cpp
The traffic light at the nearby intersection lights with a red light for 40 seconds, with yellow - 4 seconds and green - 60 seconds. Every morning at exactly 8 o'clock it  lights with a green light (traffic light in the following sequence: green, yellow, red, yellow, again: green, yellow, red, yellow and so on..). Write a program , which determines what light of the traffic light is at every second of the day.